--- modules/Exceptions.py
@@ -1,11 +1,0 @@
-class ExporterException(Exception):
-    pass
-
-
-class AtlasError(ExporterException):
-    pass
-
-
-class VcenterError(ExporterException):
-    pass
-
